Fixed incorrect inference of API URL to use as well as SSL detection
The old logic for the API url was using `hostname` which does not contain the port, the port
is present usually when going through a port forward for example `localhost:8080`. By using
`host`, if the port is present and different than standard port for protocol, it will be
present in the returned value.

The SSL detection was not working for a production build that was using a port forward.
First, the logic was that if it was not a development build, then it was necessary
using SSL. This is false when using the port forward technique.

Now, on a development build, we force to NOT use SSL. On a production build, we check if
the protocol starts with `https:`, it it's the case, then we assume SSL, otherwise, we
don't.