Fixed dmesh connection starting over again on each peer upgrade
Due to React Hooks inner working, we cannot rely on the value of the `headBlockNum`
variable coming from the `useState` above unless we have a dependency on it (which
re-triggers the effect).

However, we do not want to refresh the effect each time the head block num change.
As such, we have a local copy that we use internally and update as well as setting
the `useState` setter so the actual page is re-rendered, but without impacting the
effect (since it has a dependencies list of `[]` which force it to run only on first
render).